requirements.txt
torch>=2.0.1
torch_geometric>=2.4.0
numpy>=1.25.0
tqdm>=4.65.0
PyYAML>=6.0
networkx>=3.0
tensorboard>=2.13
